refactor(ui): extract InitialLoadingScreen + useInitialLoaderPhase
Pull the phase machine (loading→complete→fading→done) and the loading
checklist render out of AppContent into focused units:

- useInitialLoaderPhase: encapsulates both effects (transition + timers)
  and returns LoaderPhase. Easier to test in isolation; keeps AppContent
  free of timer/effect wiring.

- InitialLoadingScreen: owns the spinner, status text, and checklist
  render. Receives phase/connecting/loadingItems as props.

AppContent now calls useInitialLoaderPhase() and renders
<InitialLoadingScreen /> — no logic change.
